+++ typescript/lib/execution/metadata_providing_testsuite_service.ts
@@ -5,20 +5,23 @@ import { TestConfiguration } from "../testsuite/test_configuration"; //TODO
 import { TestSuite } from "../testsuite/test_suite"; //TODO
 import { newTestMetadata } from "../constructor_calls";
 import { ok, err, Result } from "neverthrow";
+import * as google_protobuf_empty_pb from "google-protobuf/google/protobuf/empty_pb";
+import * as grpc from "grpc";
 
 // Service handlign endpoints when the testsuite is in metadata-providing mode - i.e. NOT running a testnet, without a connection to an API container
-export class MetadataProvidingTestsuiteService implements ITestSuiteServiceServer{ //TODO - right server to implement?
+export class MetadataProvidingTestsuiteService implements ITestSuiteServiceServer { //TODO - right server to implement?
 	private readonly suite: TestSuite;
 
 	constructor(suite: TestSuite) {
 		this.suite = TestSuite;
 	}
 
-	public isAvailable(empty: null): Result<null, Error> { //TODO - should empty be null or google_protobuf_empty_pb.Empty as indicated in typescript bindings
-		return ok(empty);
+	//TODO type signature of this method in grpc bindings => isAvailable: grpc.handleUnaryCall<google_protobuf_empty_pb.Empty, google_protobuf_empty_pb.Empty>;
+	public isAvailable(): Result<grpc.handleUnaryCall<google_protobuf_empty_pb.Empty, google_protobuf_empty_pb.Empty>, Error> { //TODO - should I be using Result; how do I go about writing this method signature?
+		return ok(null);
 	}
 
-	public getTestSuiteMetadata(empty: null): Result<TestSuiteMetadata, Error> {
+	public getTestSuiteMetadata(): Result<TestSuiteMetadata, Error> { //TODO - fix method signature
 		const allTestMetadata: Map<string, TestSuiteMetadata> = new Map();
 		for (let [testName, test] of this.suite.getTests().entries()) {
 			const testConfigBuilder: TestConfigurationBuilder = new TestConfigurationBuilder();
@@ -40,15 +43,15 @@ export class MetadataProvidingTestsuiteService implements ITestSuiteServiceServe
 		return ok(testSuiteMetadata); //TODO - can there never be an error case?
 	}
 
-	public registerFiles(args: RegisterFilesArgs): Result<null, Error> { //TODO - parameter is unused
+	public registerFiles(args: RegisterFilesArgs): Result<null, Error> { //TODO - fix method signature
 		return err(new Error("Received a register files call while the testsuite service is in metadata-providing mode; this is a bug in Kurtosis"));
 	}
 
-	public setupTest(args: SetupTestArgs): Result<null, Error> { //TODO
+	public setupTest(args: SetupTestArgs): Result<null, Error> { //TODO - fix method signature
 		return err(new Error("Received a setup test call while the testsuite service is in metadata-providing mode; this is a bug in Kurtosis"));
 	}
 
-	public runTest(args: RunTestArgs): Result<null, Error> { //TODO
+	public runTest(args: RunTestArgs): Result<null, Error> { //TODO - fix method signature
 		return err(new Error("Received a run test call while the testsuite service is in metadata-providing mode; this is a bug in Kurtosis"));
 	}
 }