+++ typescript/lib/execution/test_suite_executor.ts
@@ -1,23 +1,23 @@
 import { ITestSuiteServiceServer } from "../../kurtosis_testsuite_rpc_api_bindings/testsuite_service_grpc_pb";
-import { NetworkContext, newApiContainerServiceClient } from "kurtosis-js-lib"; //TODO
+import { NetworkContext, newApiContainerServiceClient } from "kurtosis-core-api-lib"; //TODO
 import { TestExecutingTestsuiteService } from "./test_executing_testsuite_service";
 import { TestSuiteConfigurator } from "./test_suite_configurator";
 import { MetadataProvidingTestsuiteService } from "./metadata_providing_testsuite_service";
 import { TestSuite } from "../testsuite/test_suite"; //TODO
 import { KurtosisTestsuiteDockerEnvVar, ENCLAVE_DATA_VOLUME_MOUNTPOINT } from "../../kurtosis_testsuite_docker_api/kurtosis_testsuite_docker_api";
-import { } from "../../kurtosis_testsuite_rpc_api_consts/kurtosis_testsuite_rpc_api_consts";
+//import { LISTEN_PORT, LISTEN_PROTOCOL } from "../../kurtosis_testsuite_rpc_api_consts/kurtosis_testsuite_rpc_api_consts"; //need for minimial_grpc_server
+//"github.com/kurtosis-tech/minimal-grpc-server/server" //TODO
 import { Result, err, ok } from "neverthrow";
-"github.com/kurtosis-tech/minimal-grpc-server/server" //TODO
 import * as grpc from "grpc";
-"os"
+
 //TODO Below
-//import { process } from "node"; //TODO - don't need line but need the dependency = npm i --save-dev @types/node
-import * as dotenv from 'dotenv';
+//import { process } from "node"; //TODO - don't need line but need the dependency = npm i --save-dev @types/node for EnvVar
+import * as dotenv from 'dotenv'; //"os" //TODO
 //import * as time from "date-and-time"; //npm i --save-dev @types/date-and-time
 //import * as date from "date-fns"; //npm install date-fns --save
 
 dotenv.config();
-const GRPC_SERVER_STOP_GRACE_PERIOD: number = 5 * date.getSeconds(); //TODO - how does time work in typescript
+//const GRPC_SERVER_STOP_GRACE_PERIOD: number = 5 * time.Second; //TODO - needed for minimial GRPC server
 
 class TestSuiteExecutor {
 	private readonly configurator: TestSuiteConfigurator;
@@ -86,11 +86,11 @@ class TestSuiteExecutor {
             kurtosis_testsuite_rpc_api_bindings.RegisterTestSuiteServiceServer(grpcServer, testsuiteService) //TODO - RegisterTestSuiteServiceServer (no such line)
         }
 
-        //TODO - how to call minimial grpc server (how would I call this minimial grpc server)
+        //TODO TODO TODO - don't need this right now, but needed when we hit lambda-api-lib
         // testsuiteServer := server.NewMinimalGRPCServer(
         //     kurtosis_testsuite_rpc_api_consts.ListenPort,
         //     kurtosis_testsuite_rpc_api_consts.ListenProtocol,
-        //     grpcServerStopGracePeriod,
+        //     GRPC_SERVER_STOP_GRACE_PERIOD,
         //     []func(desc *grpc.Server) {
         //         testsuiteServiceRegistrationFunc,
         //     },
@@ -98,92 +98,6 @@ class TestSuiteExecutor {
         // if err := testsuiteServer.Run(); err != nil {
         //     return stacktrace.Propagate(err, "An error occurred running the testsuite server")
         // }
-
-
-        // TODO TODO TODO - I don't if I need to write this up in typescript or not, but here is the go code for reference
-        // Might delete if not needed
-
-        // import (
-        //     "fmt"
-        //     "github.com/palantir/stacktrace"
-        //     "github.com/sirupsen/logrus"
-        //     "google.golang.org/grpc"
-        //     "net"
-        //     "os"
-        //     "os/signal"
-        //     "syscall"
-        //     "time"
-        // )
-
-        // type MinimalGRPCServer struct {
-        //     listenPort uint32
-        //     listenProtocol string
-        //     stopGracePeriod time.Duration  // How long we'll give the server to stop after asking nicely before we kill it
-        //     serviceRegistrationFuncs []func(*grpc.Server)
-        // }
-
-        // // Creates a minimal gRPC server but doesn't start it
-        // // The service registration funcs will be applied, in order, to register services with the underlying gRPC server object
-        // func NewMinimalGRPCServer(listenPort uint32, listenProtocol string, stopGracePeriod time.Duration, serviceRegistrationFuncs []func(*grpc.Server)) *MinimalGRPCServer {
-        //     return &MinimalGRPCServer{listenPort: listenPort, listenProtocol: listenProtocol, stopGracePeriod: stopGracePeriod, serviceRegistrationFuncs: serviceRegistrationFuncs}
-        // }
-
-        // func (server MinimalGRPCServer) Run() error {
-        //     grpcServer := grpc.NewServer()
-
-        //     for _, registrationFunc := range server.serviceRegistrationFuncs {
-        //         registrationFunc(grpcServer)
-        //     }
-
-        //     listenAddressStr := fmt.Sprintf(":%v", server.listenPort)
-        //     listener, err := net.Listen(server.listenProtocol, listenAddressStr)
-        //     if err != nil {
-        //         return stacktrace.Propagate(
-        //             err,
-        //             "An error occurred creating the listener on %v/%v",
-        //             server.listenProtocol,
-        //             server.listenPort,
-        //         )
-        //     }
-
-        //     // Signals are used to interrupt the server, so we catch them here
-        //     termSignalChan := make(chan os.Signal, 1)
-        //     signal.Notify(termSignalChan, syscall.SIGINT, syscall.SIGTERM, syscall.SIGQUIT)
-
-        //     grpcServerResultChan := make(chan error)
-
-        //     go func() {
-        //         var resultErr error = nil
-        //         if err := grpcServer.Serve(listener); err != nil {
-        //             resultErr = stacktrace.Propagate(err, "The gRPC server exited with an error")
-        //         }
-        //         grpcServerResultChan <- resultErr
-        //     }()
-
-        //     // Wait until we get a shutdown signal
-        //     <- termSignalChan
-
-        //     serverStoppedChan := make(chan interface{})
-        //     go func() {
-        //         grpcServer.GracefulStop()
-        //         serverStoppedChan <- nil
-        //     }()
-        //     select {
-        //     case <- serverStoppedChan:
-        //         logrus.Debug("gRPC server has exited gracefully")
-        //     case <- time.After(server.stopGracePeriod):
-        //         logrus.Warnf("gRPC server failed to stop gracefully after %v; hard-stopping now...", server.stopGracePeriod)
-        //         grpcServer.Stop()
-        //         logrus.Debug("gRPC server was forcefully stopped")
-        //     }
-        //     if err := <- grpcServerResultChan; err != nil {
-        //         // Technically this doesn't need to be an error, but we make it so to fail loudly
-        //         return stacktrace.Propagate(err, "gRPC server returned an error after it was done serving")
-        //     }
-
-	    // return nil
-        //}
-
 
         return ok(null);
     }